Started work on letting buildings be built.
Added "Buildable", an Actor Component that contains in itself the required resources and amounts necessary to build a building, as well as associated functions. I've tried very hard to keep it confined exclusively to building the building, and as such it only handles:

* Does the building have all of the necessary resources other than Labor?
* Does the building have all of the necessary resources, including labor?
* Costs and 'Currently Delivered Resources'
* Is the building completed?

Also have begin to build the "ResourceHolder" Actor Component, which will be how we track resources, unless I figure out a better place to put it, I wasn't able to find a more 'abstract' thing to use for it.

To support this, I have also added "MapFunctions", a function library which will contain general purpose functions used to mess with Maps, so we can avoid having to replicate a bunch of code unecessarily.