Resource transfer and storage (mostly) completed
TL;DR USE THE "TRY" SERIES OF FUNCTIONS RATHER THAN DIRECTLY EDITING VALUES.

"ResourceFunctions" contains all* of the necessary code to transfer resources from one Actor with the 'Resource Holder' Actor Component to another.

It's possible to 'spend' resources from any Actor's "Resource Map" as long as you pass a reference to that Actor, but in general try to use the "Try" series of functions, i.e.:

"Resource Action: Try Transfer Resource(s)"
"Try to Spend Resource"
"Try to Buy with Resources"

...as these functions incorporate error checking and validation to make sure that nothing explodes, and (mostly) will return error messages to indicate what didn't work.

The simpler functions like "rMap Increment Resource Value" and "rMap Decrement Resource Value" DO NOT contain code to see if you actually have enough, and are not meant to be called as the primary means to spend resources.