Added "Take Map from Target"
This function functions (haha) as a wrapper for the "Try Transfer Resources" function, causing it to respect "Taker" inventory limits as well as when the "Giver" has less than the desired quantity.

Time taken to make: 1 hour, time taken to go crazy looking for the cause of weird behavior, add a ton of debug functionality, add in more logic to handle the edge cases which were causing the bug, and then mutter darkly about it in the update push: 3 hours.