Started adding peon build logic
So far they have a "Deliver Map", a "destination Actor reference", and a bool for "am I collecting resources in order to deliver them?"

Next up is probably making some arrays to keep segregated lists of buildings which can then be looked at by Peon logic for behaviors. And then we can finally test it all.