Added "is My Inventory Full" +"LevelUp" Component + Misc additions
"is My Inventory Full" takes in an Actor reference, determines of that Actor's "Resource Map" is full or not, then returns that value. If the passed Actor reference happens to be a BP_Peon, it will set the "Inventory Full?" flag on that Peon.

The function includes error checking and debug output.

New Actor Component which contains all of the information necessary for the host Combat Actor to gain levels. All of the math should work, but it does not yet have an effect on stats.