Finished hooking up LevelUp granting stats
Leveling up now grants stats.  There's no implemented way to actually gain xp but the functions which do it are completed.  There is also a toggle for, "manual level up only", intended to be used so that dragons do not level by themselves and instead will have a custom wrapper function for the levelup process which deactivates the flag and then reactivates it after they level.