Started work on the Weapon actor component.
Not done yet, but started to lay the groundwork. One additional functionality being added to 'hits', recursion. The recursion value is an integer and it will reapply the same 'hit', useful for things like shotgun blasts or anything that has a larger number of the same kind of 'hit' which would be cumbersome to do with a map.